Created new java class named ActionSetContextCommon and removed duplicate code in 2 files namely UMLActivityGraphContextComboBoxModel and UMLStateMachineContextComboBoxModel
