+++ src/argouml-core-umlpropertypanels/src/org/argouml/core/propertypanels/ui/UMLActivityGraphContextComboBoxModel.java
@@ -114,40 +114,14 @@ class UMLActivityGraphContextComboBoxModel extends  UMLComboBoxModel {
     public void modelChange(UmlChangeEvent evt) {
         /* Do nothing by design. */
     }
-    
+
+    /**
+     * Returns a shared action to set the context of the ActivityGraph.
+     *
+     * Uses the common ActionSetContextCommon to avoid code duplication.
+     */
     public Action getAction() {
-        return new ActionSetContext();
+        return new ActionSetContextCommon();
     }
-    
-    private class ActionSetContext extends UndoableAction {
 
-        /**
-         * The UID.
-         */
-        private static final long serialVersionUID = -8118983979324112900L;
-
-        /**
-         * Constructor for ActionSetCompositeStateConcurrent.
-         */
-        protected ActionSetContext() {
-            super(Translator.localize("action.set"), null);
-            // Set the tooltip string:
-            putValue(Action.SHORT_DESCRIPTION, 
-                    Translator.localize("action.set"));
-        }
-
-        /*
-         * @see java.awt.event.ActionListener#actionPerformed(java.awt.event.ActionEvent)
-         */
-        public void actionPerformed(ActionEvent e) {
-            super.actionPerformed(e);
-            UMLComboBox source = (UMLComboBox) e.getSource();
-            Object target = source.getTarget();
-            if (Model.getFacade().getContext(target)
-                    != source.getSelectedItem()) {
-                Model.getStateMachinesHelper().setContext(
-                        target, source.getSelectedItem());
-            }
-        }
-    }
 }
