+++ src/argouml-core-umlpropertypanels/src/org/argouml/core/propertypanels/ui/UMLStateMachineContextComboBoxModel.java
@@ -109,40 +109,14 @@ class UMLStateMachineContextComboBoxModel
     public void modelChanged(UmlChangeEvent evt) {
         /* Do nothing by design. */
     }
-    
+
+    /**
+     * Returns a shared action to set the context of the StateMachine.
+     *
+     * Uses the common ActionSetContextCommon to avoid code duplication.
+     */
     public Action getAction() {
-        return new ActionSetContextStateMachine();
+        return new ActionSetContextCommon();
     }
-    
-    private class ActionSetContextStateMachine extends UndoableAction {
 
-        /**
-         * The UID.
-         */
-        private static final long serialVersionUID = -8118983979324112900L;
-
-        /**
-         * Constructor for ActionSetCompositeStateConcurrent.
-         */
-        protected ActionSetContextStateMachine() {
-            super(Translator.localize("action.set"), null);
-            // Set the tooltip string:
-            putValue(Action.SHORT_DESCRIPTION, 
-                    Translator.localize("action.set"));
-        }
-
-        /*
-         * @see java.awt.event.ActionListener#actionPerformed(java.awt.event.ActionEvent)
-         */
-        public void actionPerformed(ActionEvent e) {
-            super.actionPerformed(e);
-            UMLComboBox source = (UMLComboBox) e.getSource();
-            Object target = source.getTarget();
-            if (Model.getFacade().getContext(target)
-                    != source.getSelectedItem()) {
-                Model.getStateMachinesHelper().setContext(
-                        target, source.getSelectedItem());
-            }
-        }
-    }
 }
